NEW: DataLoafPackageDirFmt (#4)
This adds three new formats:

**DataLoafPackageDirFmt**
**DataPackageSchemaFileFormat**
**FrictionlessCSVFileFormat**

The `DataLoafPackageDirFmt` is intended for artifacts containing multiple files of the same size (number of columns & rows). The analogy here is a 'loaf' of data containing identical data 'slices (i.e. files) with each file in the form of a `FrictionlessCSVFileFormat`. The loaf also contains 'nutrition facts' (metadata for the data slices) in the form of a `DataPackageSchemaFileFormat`.